backend/pyproject.toml
[project]
name = "df-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "fastapi[standard]>=0.115.8",
    "pillow>=11.1.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
